bindings/go: safe finalization
When calling a function on a struct with a finalizer, the struct can
be destroyed/finalized within the function if it will no longer be
reachable. Example: https://play.golang.org/p/Lv19twqhDEU

This is fine in the example, but imagine if `data` were instead
a cgo pointer that gets freed by the finalizer. We'd end up
using the pointer after freeing it and badness would ensue.

Instead, we should ensure that the struct stays alive within any
functions in which its wrapped pointer is used. We can use
`defer runtime.KeepAlive()` to this end.

I looked for all structures that have a finalizer set and audited
their `.ptr` usage, adding `runtime.KeepAlive()`s as necessary.
This is probably a bit overkill, as the struct will stay alive
beyond most of those function calls, but IMO it's better to be
explicit here.